etl/pyproject.toml
[tool.poetry]
name = "extract"
version = "0.1.0"
description = ""
authors = ["Morgan Brooks <morgan@urbanlogiq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
pyairtable = "^2.3.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
